Move types from types.go into different files for separation of concerns.
