+++ cmd/discovery/connector_kubernetes.go
@@ -3,11 +3,11 @@
 package main
 
 import (
-	"github.com/jaku01/caddyservicediscovery/internal/caddy"
-	"github.com/jaku01/caddyservicediscovery/internal/kubernetes"
+	"github.com/jaku01/caddyservicediscovery/internal/provider"
+	"github.com/jaku01/caddyservicediscovery/internal/provider/kubernetes"
 )
 
 // Returns a Kubernetes connector when built with the 'kubernetes' tag.
-func newServiceDiscoveryProviderConnector() (caddy.ServiceDiscoveryProvider, error) {
+func newServiceDiscoveryProviderConnector() (provider.ServiceDiscoveryProvider, error) {
 	return kubernetes.NewKubernetesConnector()
 }
